introduce zoom feature. Address #3
This commit introduces zoom to the AppsGate.debugger. Users can now zoom in and out using the focusline.
When zooming in, more detailed information can be requested from the server to refresh the timelines.
This automatic reload can be enabled or disabled: when disabled the zoom will be equivalent to a numerical zoom.